==> 1_rt.Rmd <==
---
title: "Estimated Rt"
output: html_notebook
---

## Load packages
```{r}
library(tidyverse)
library(EpiEstim)
```

## Load data
```{r}
regions <- c('Peel Public Health', 'Halton Region Health Department', "Toronto Public Health", "York Region Public Health Services", "Durham Region Health Department")
data <- read_csv('data/conposcovidloc.csv') %>% 
  filter(Reporting_PHU %in% regions) %>% 
  group_by(Reporting_PHU, Case_Reported_Date) %>% 
  count() %>% 
  rename(Region=Reporting_PHU, Date=Case_Reported_Date) %>% 
  mutate(Region = case_when(Region == "Durham Region Health Department" ~ 'Durham',
                            Region == 'Halton Region Health Department' ~ 'Halton',
                            Region == 'Peel Public Health' ~ 'Peel',
                            Region == "Toronto Public Health" ~ 'Toronto',
                            Region == "York Region Public Health Services" ~ 'York')) %>% 
  pivot_wider(names_from = Region, values_from = n, values_fill = 0)

write_csv(data, 'data/cleaned_cases.csv')
```

## Summary of data
```{r}
summary(data)

rt_data <- data %>% 
  filter(Date >= '2020-03-16')
```

## Make df's for R calculation
```{r}
toronto <- rt_data %>% 
  select(Date, Toronto) %>% 
  rename(dates=Date, I=Toronto)
peel <- rt_data %>% 
  select(Date, Peel) %>% 
  rename(dates=Date, I=Peel)
york <- rt_data %>% 
  select(Date, York) %>% 
  rename(dates=Date, I=York)
durham <- rt_data %>% 
  select(Date, Durham) %>% 
  rename(dates=Date, I=Durham)
halton <- rt_data %>% 
  select(Date, Halton) %>% 
  rename(dates=Date, I=Halton)
```

## Calculate Rt
```{r}
tr <- estimate_R(toronto, method = 'parametric_si',
           config = make_config(list(mean_si = 5.29, std_si = 5.32)))
pr <- estimate_R(peel, method = 'parametric_si',
           config = make_config(list(mean_si = 5.29, std_si = 5.32)))
yr <- estimate_R(york, method = 'parametric_si',
           config = make_config(list(mean_si = 5.29, std_si = 5.32)))
dr <- estimate_R(durham, method = 'parametric_si',
           config = make_config(list(mean_si = 5.29, std_si = 5.32)))
hr <- estimate_R(halton, method = 'parametric_si',
           config = make_config(list(mean_si = 5.29, std_si = 5.32)))
```

## Toronto
```{r}
plot(tr, 'R')
```

## Peel
```{r}
plot(pr, 'R')
```

## York
```{r}
plot(yr, 'R')
```

## Durham
```{r}
plot(dr, 'R')
```

## Halton
```{r}
plot(hr, 'R')
```

```{r, include=FALSE}
# Save plots

png('estR/Toronto_R.png', 
    width=11,
    height=8.5,
    units='in',
    res = 144)
plot(tr, 'R')
dev.off()

png('estR/Peel_R.png', 
    width=11,
    height=8.5,
    units='in',
    res = 144)
plot(pr, 'R')
dev.off()

png('estR/York_R.png', 
    width=11,
    height=8.5,
    units='in',
    res = 144)
plot(yr, 'R')
dev.off()

png('estR/Durham_R.png', 
    width=11,
    height=8.5,
    units='in',
    res = 144)
plot(dr, 'R')
dev.off()

png('estR/Halton_R.png', 
    width=11,
    height=8.5,
    units='in',
    res = 144)
plot(hr, 'R')
dev.off()
```

## Get data in usable form
```{r}
tr <- tr$R %>% 
  add_column(date_start = tr$dates[tr$R$t_start],
             date_end = tr$dates[tr$R$t_end]) %>% 
  add_column(Region = rep('Toronto', times=nrow(tr$R)))
pr <- pr$R %>% 
  add_column(date_start = pr$dates[pr$R$t_start],
             date_end = pr$dates[pr$R$t_end]) %>% 
  add_column(Region = rep('Peel', times=nrow(pr$R)))
yr <- yr$R %>% 
  add_column(date_start = yr$dates[yr$R$t_start],
             date_end = yr$dates[yr$R$t_end]) %>% 
  add_column(Region = rep('York', times=nrow(yr$R)))
dr <- dr$R %>% 
  add_column(date_start = dr$dates[dr$R$t_start],
             date_end = dr$dates[dr$R$t_end]) %>% 
  add_column(Region = rep('Durham', times=nrow(dr$R)))
hr <- hr$R %>% 
  add_column(date_start = hr$dates[hr$R$t_start],
             date_end = hr$dates[hr$R$t_end]) %>% 
  add_column(Region = rep('Halton', times=nrow(hr$R)))

```



## Write csv's
```{r}
estR <- rbind(tr, pr, yr, dr, hr)
write_csv(estR, 'estR/estR.csv')
```
